pmb.core.pkgrepo:pkgrepo_iter_package_dirs: speed up finding pkg dirs
This uses os.walk instead of glob to have more control over which
directories are traversed when searching for package dirs.
It skips (hidden) directories that start with '.', and any directories
that are subdirectories of package dirs (I don't think this is supported
by abuild?)

It fixes a long delay, for example on my current pmaports dir:

[13:22:39] $ pmbootstrap /home/clayton/src/pmbootstrap/pmbootstrap.py --details-to-stdout build postmarketos-initramfs
[13:22:39] Pmbootstrap v3.5.2 (Python 3.12.11 (main, Jun  4 2025, 09:54:34) [GCC 14.3.0])
[13:22:45] % cd /home/clayton/src/pmaports; git remote -v
(yikes! 6 seconds!)
[13:22:45] % cd /home/clayton/src/pmaports; git show origin/master:channels.cfg
[13:22:45] Attempting to build: postmarketos-initramfs

I've seen this delay be upwards of 30 seconds or more in some cases.

With this change, there's no observable delay when searching for
packages in my local pmaports directory.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2648