pmb.build.kconfig: add a "generate" cmd to build kconfig from fragments
This adds a new command, "generate" to pmbootstrap kconfig to generate
a full kernel config from various fragments and validate the generated
config so that folks can be confident that options they expect to be in
the final config are actually there and not silently dropped by the
kernel build system. The aim is to make it easier to maintain and deploy
distro-wide kconfig, and for contributors to have a much smaller and
more manageable set of kconfig to review/maintain for kernel packages.

This new command generates a full kernel config from:

1) A defconfig in the kernel source if one is specified in the APKBUILD

2) A config fragment created from kconfigcheck.toml for the specific
   kernel/arch, and copies it to the kernel pkg aports dir for source
   control

3) Any fragments it finds in the kernel's aports dir

Once the full config is generated from these fragments, then:

1) It's copied to the kernel's pmaports dir

2) `kconfig check` is run on it to make sure it still passes the
   kconfigcheck.toml

3) It validates that options enabled by fragments found in the aports
   dir are also set as expected in the generated config

Kernel pkg maintainers can then checksum/commit the full config into
pmaports and use it to build the kernel. The idea is that any kconfig
changes would happen in kconfigcheck.toml and/or in fragments in kernel
pkg dirs, and maintainers would run `pmb kconfig generate` to update the
full config in pmaports, commit, then build from it.

I tried to re-use as many mechanisms in pmbootstrap as possible.
I felt like using kconfigcheck.toml as the "one source" for distro
kconfig requirements was better than previous approaches where we tried
generating new fragments and introducing new logic (e.g. jinja) to
apply config conditionally. A lot of the bits for this are already in
pmbootstrap, AND this let's us generate a fragment from the same thing
that we use later to validate the final config, without having to make
further changes to pmb.

This change should have no impact at all on kernels in pmaports where
maintainers are editing them 'by hand' or with kconfig edit. Maintainers
can continue to do what we've done all along if they want, and kconfig
edit/check and so on should function the same.

Signed-off-by: Clayton Craft <craftyguy@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2625>