pmb.parse.deviceinfo: Handle deviceinfo_initfs_compression
Maybe this is a bit of an overkill implementation since we don't
actually use the compression level for anything, but maybe it'll be
useful in the future, I don't know.

Additionally, while the deviceinfo reference says that the format is
FORMAT:LEVEL, in practice it seems both FORMAT and FORMAT:LEVEL are
allowed, so handle both cases.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2645>