Honor version constraints in makedepends
systemd-stage0 has dbus-dev<9990 in its makedepends to force that the
Alpine package gets used instead of the pmaports dbus package. Without
this patch, the "<9990" part gets ignored and pmbootstrap tries to use
the pmaports dbus package during the build, which depends on systemd and
therefore breaks bootstrapping systemd for a new architecture / stable
branch.

Fix this by handling the >=, >, <=, < operators in such makedepends. If
the version constraints don't match the pmaports package, then don't use
it for the dependency tree.

Fixes: issue 2663
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2695>