kconfig: Reintroduce command line argument to check for categories
Commit 834f0ecf ("kconfig check: remove --kconfig-check-{name} args")
has removed the ability to check for specific extra categories.

Reintroduce this support by allowing the user to provide a list of
categories to check for. For kconfigs from packages, this is in addition
to the ones defined in the options in the APKBUILDs.

$ pmbootstrap kconfig check linux-postmarketos-qcom-sc7280 --categories uefi

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2697>