.gitignore: Add mypyc section
When building pmbootstrap with mypyc, this directory is created. While
it isn't possible to successfully build pmbootstrap with mypyc yet, I
experiment with it every once in a while and it is nice to have this
folder ignored for that reason.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2718>