pmb: Make command timeout opt-in
See linked issues for full rationale, but in summary this causes more
problems than it solves as it is a non-obvious feature that is easy to
forget about and as such easily can waste your time in that it can kill
long-running processes that you'll likely just have to restart again
except with the timeout set higher.

The original reasoning for having this was to prevent CI and other
builders from getting stuck wasting CPU time—which is reasonable—but not
something that affects the average user trying to use pmbootstrap on
their own system to install postmarketOS or build packages.

As such, make this an opt-in feature rather than an opt-out one.

Closes https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/issues/2676
Closes https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/issues/2132
Closes https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/issues/2363
Closes https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/issues/1798

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2724>