flasher/frontend.py: fix "The recovery zip has not been generated yet"
After sucessfully generating an twrp flashable .zip as proposed e.g. on
the samsung-dream (Samsung S8) wiki page:

  pmbootstrap install --android-recovery-zip \
                      --recovery-install-partition=data --password 1234

the generated .zip can not be flashed with the command

  pmbootstrap flasher --method=adb sideload

but results in the error messages:

  The recovery zip has not been generated yet, please run
  'pmbootstrap install' with the '--android-recovery-zip' parameter
  first!

This may be caused by the mountpoint not having the expected name as
well as the concatenation of PosixPaths via the "/" operator ignoring
the first parameter if the second starts with a leading "/".

Furthermore, the existance check should not be performed relative to
native chroot, but acessible by the python script of pmbootstrap.

Tweaked-by: Oliver Smith <ollieparanoid@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2737>