pmb.config.init: Automatically switch to Edge if creating a new port
All new packages, including ones for device ports, should go through
Edge. Previously pmbootstrap didn't really enforce this but it was also
pretty easy to overlook because there were no immediate problems with
it. Now that we split off downstream ports into their own "downstream"
device category, suddenly this causes crashes as this downstream
category is not present on stable branches. So, if you try to make a
downstream port on a stable branch, you get a crash because pmbootstrap
tries to operate in a nonexistent directory.

To resolve this, just automatically switch to the development branch
(i.e. Edge) if a user selects a stable branch. Initially I just had
pmbootstap exit and display an error, but this seems like less friction
and confusion for users, which is particularly important in this context
as it may be one of their first encounters with pmbootstrap.

Closes https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/issues/2677

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2738>