pmb.core.context: Re-enable ccache by default
After the pmbv3 rewrite, caching was disabled by default with no way to
turn it back on. pmbootstrap offers a `--no-ccache` option, but there
is no `--use-ccache` option, meaning that pmbootstrap should default to
using ccache.

There is a FIXME in the Context class that it is in, but that looks
like it's aimed to pull out the caching behavior into its own class.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2740>