pmb.parse: Don't prefer exact package match over other providers
When a package depends on e.g. systemd-dev and dbus-dev, pmbootstrap
would install systemd-dev and dbus-dev, but then also look at the
dependencies of dbus-dev, which depends on systemd-stage0-dev, and then
add systemd-stage0-dev to the list of packages to install, resulting in
a conflict.

It is correct to add systemd-stage0-dev to the list of packages to
install, but pmbootstrap chose the wrong provider - there are two:
systemd-stage0-dev with origin systemd-stage0 and systemd-dev with
origin systemd. The current codepath that forces installing a package
with the exact name over any other provider is therefore wrong.

Signed-off-by: Jens Reidel <adrian@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2747>