pmb: Fix kernel architecture logic
This is a bit complicated. The directory for x86_64 and x86 in the
kernel is x86, but ARCH=x86_64 will default to x86_64_defconfig and
ARCH=i386 will default to i386_defconfig (there is no x86_defconfig).

When running make defconfig on x86 or x86_64, the kernel uses uname -m to
determine which of the two defconfigs to use. In cross-native on x86_64
when targeting x86, this returns x86_64 and ends up in the wrong config.

We can work around this by passing ARCH=i386 when targeting x86 and
ARCH=x86_64 when targeting x86_64.

Signed-off-by: Jens Reidel <adrian@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2751>