pmb.build.kconfig: Hardcode native chroot even more
The rest of the logic hardcodes the native chroot in a lot of places,
but not here. When cross-native is enabled, the native chroot is used in
all cases... except when cross compile is determined to be unnecessary
(e.g. when targeting x86 on x86_64).

That completely messed with the rest of the logic here and I think this
is the less intrusive fix (the alternative being changing unnecessary
cross compiles to use the native chroot, which sounds like a bad idea
considering it'll change uname -m output unexpectedly).

Signed-off-by: Jens Reidel <adrian@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2751>